Large refactor: centralize all run types, and add completion to admin commands
Instead of having hacky references to "p", "c", etc all over the place,
we now have a central location (in Agronet) to define all run types.

Citadel now uses this information everywhere, notably as part of
player inventory management and item generation.

Relates to #51

+++ src/main/kotlin/org/trackedout/citadel/mongo/MongoPlayer.kt
@@ -22,6 +22,6 @@ data class LastLocation(
     @BsonId val id: ObjectId,
 
     val x: Double,
-    val y: Long,
+    val y: Double,
     val z: Double,
 )
